feat: 링크 입력창 UI개선, 링크 입력 모달 삭제 링크 입력창 UI개선, 링크 입력 모달 삭제 (#111)
