+++ frontEnd/src/components/common/SectionWrapper.tsx
@@ -1,5 +1,5 @@
 import { ReactNode } from 'react';
 
 export default function Section({ children }: { children: ReactNode }) {
-  return <div className="w-full h-full border rounded-lg drop-shadow-lg">{children}</div>;
+  return <div className="w-full h-full border rounded-lg drop-shadow-lg bg-primary">{children}</div>;
 }
